added money_parser to case_parser
